# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Moskollektor Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/work-orders": {
      "post": {
        "tags": [
          "work-orders"
        ],
        "summary": "Post Work Order",
        "description": "Create a draft work order.\n\nArgs:\n    body: The creation request; mode must be \"draft\".\n    user: The authenticated caller, injected by require_any_permission.\n\nReturns:\n    The created work order.\n\nRaises:\n    ApiError: 422 for mode != \"draft\" or an unknown work_type/priority/\n        target_entity_type; 403 if facility_id is outside the caller's scope.",
        "operationId": "post_work_order_api_v1_work_orders_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "work-orders"
        ],
        "summary": "Get Work Orders",
        "description": "List work orders within the caller's scope, newest first.\n\nArgs:\n    facility_id: Optional exact facility filter.\n    status: Optional exact status filter.\n    work_type: Optional exact WORK_TYPES filter.\n    cursor: Opaque pagination cursor.\n    limit: Page size, 1-200.\n    user: The authenticated caller, injected by require_permission.\n\nReturns:\n    A WorkOrderListEnvelope containing only work orders the caller may see.",
        "operationId": "get_work_orders_api_v1_work_orders_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "work_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Work Type"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkOrderListEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWorkOrderRequest": {
        "properties": {
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "source_risk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Risk Id"
          },
          "facility_id": {
            "type": "string",
            "title": "Facility Id"
          },
          "target_entity_type": {
            "type": "string",
            "title": "Target Entity Type"
          },
          "target_entity_id": {
            "type": "string",
            "title": "Target Entity Id"
          },
          "work_type": {
            "type": "string",
            "title": "Work Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "mode",
          "facility_id",
          "target_entity_type",
          "target_entity_id",
          "work_type",
          "priority",
          "due_at",
          "description"
        ],
        "title": "CreateWorkOrderRequest",
        "description": "Body for POST /api/v1/work-orders.\n\nmode must be \"draft\" -- the only mode this backend ever supports, since\nno real external work-order system exists to submit to."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkOrderListEnvelope": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/WorkOrderOut"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/WorkOrderListMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "WorkOrderListEnvelope",
        "description": "The standard {data, meta} envelope for GET /api/v1/work-orders."
      },
      "WorkOrderListMeta": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "next_cursor",
          "total",
          "generated_at"
        ],
        "title": "WorkOrderListMeta",
        "description": "Cursor-pagination metadata for the work-order list envelope."
      },
      "WorkOrderOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_number": {
            "type": "string",
            "title": "Display Number"
          },
          "source_risk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Risk Id"
          },
          "facility_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Facility Id"
          },
          "target_entity_type": {
            "type": "string",
            "title": "Target Entity Type"
          },
          "target_entity_id": {
            "type": "string",
            "title": "Target Entity Id"
          },
          "work_type": {
            "type": "string",
            "title": "Work Type"
          },
          "priority": {
            "type": "string",
            "title": "Priority"
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_number",
          "source_risk_id",
          "facility_id",
          "target_entity_type",
          "target_entity_id",
          "work_type",
          "priority",
          "due_at",
          "description",
          "comment",
          "status",
          "created_by",
          "created_at",
          "updated_at",
          "version"
        ],
        "title": "WorkOrderOut",
        "description": "One work order's full representation."
      },
...
    }
  }
}